Moved eventResolver
