+++ src/main/scala/com/github/emildafinov/ad/sdk/event/resolvers/EventResolverFactory.scala
@@ -1,11 +1,9 @@
 package com.github.emildafinov.ad.sdk.event.resolvers
 
-import com.github.emildafinov.ad.sdk.{ClientDefinedCredentialsModule, ConnectorResolversAkkaDependenciesModule}
 import com.github.emildafinov.ad.sdk.authentication.AppMarketCredentialsSupplier
 import com.github.emildafinov.ad.sdk.event.EventResolverFactory
 import com.github.emildafinov.ad.sdk.event.marshallers.EventResultMarshallersModule
-import com.github.emildafinov.ad.sdk.event.payloads.EventResolver
-import com.github.emildafinov.ad.sdk.event.responses.SubscriptionOrderResponse
+import com.github.emildafinov.ad.sdk.{ClientDefinedCredentialsModule, ConnectorResolversAkkaDependenciesModule}
 
 object EventResolverFactory {
   def apply(credSupplier: AppMarketCredentialsSupplier): EventResolverFactory = {
@@ -17,9 +15,7 @@ object EventResolverFactory {
       override protected val credentialsSupplier: AppMarketCredentialsSupplier = credSupplier
     }
 
-    new EventResolverFactory {
-      override def subscriptionOrderResolver(): EventResolver[SubscriptionOrderResponse] = factoryApplicationContext.subscriptionOrderEventResolver
-    }
+    () => factoryApplicationContext.subscriptionOrderEventResolver
   }
 }
